feat: Refactor Hero component to improve search functionality
Refactor the Hero component to improve the search functionality. This includes adding the ability to submit the search query by pressing the Enter key and updating the search button to be more accessible. These changes enhance the user experience by providing a more intuitive and efficient way to perform searches on the website.