feat: Add functionality to add medical folders
This commit adds the functionality to add medical folders in the MedicalFolders component. The handleAddFolder function is implemented to make a POST request to the server and add a new folder with the provided name. The user is prompted to enter a name, and if no name is entered, an error message is displayed. The loading state is managed during the API call, and the folders state is updated with the response data upon success. Error handling is also implemented to display an error message if the API call fails. This feature improves the user experience by allowing them to organize their medical documents into folders.